requirements.txt
astroid==3.3.10
certifi==2025.6.15
charset-normalizer==3.4.2
defusedxml==0.7.1
dill==0.4.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fonttools==4.58.5
fpdf2==2.8.3
freezegun==1.5.3
idna==3.10
iniconfig==2.1.0
isort==6.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.4
sendgrid==6.12.4
six==1.17.0
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
